[gldemo] factor out gltiles module
